Rename _errors.py to errors.py.
